refactor [main.js, constants.js, cards.js]: adds functions, exports
