+++ js/main.js
@@ -1,1 +1,5 @@
-import './thumbnails.js';
+import { createPublications } from './mock-publications.js';
+import { renderCards } from './cards.js';
+
+const publications = createPublications();
+renderCards(publications);
